Moved initialize out to separate file
And made the counter atomic so multiple goroutines will be able to call
it
